==> notebooks/1.0-vb-initial-purplesharp-data-exploration.ipynb <==
import pandas as pd
from pandas.io import json
import datetime
import time

path = "../data/raw/purplesharp_ad_playbook_I_2020-10-22042947.json"
df = json.read_json(path, lines=True)

len(df)

#### Kerberos Attack Related Features Explained (See paper on row 8 on google Sheets)  

AuthenticationPackageName: Indicates the authentication type. This could, for example, be Kerberos or NTLM.  
ElevatedToken: Whether or not the logon has elevated administrator privileges.  
ImpersonationLevel: The level of impersonation the logon is allowed to. Most common is to have impersonation privileges allowing to impersonate the client on the local system, but not remote systems.  
IpAddress: The source IP address where the logon originated from.  
IpPort: The source IP port of the remote system.  
KeyLength: The key length in bits for an NTLM authenticated logon.  
LmPackageName: The package name of an NTLM authenticated logon.  
LogonProcessName: The trusted logon process registered with the local security authority handling the logon.  
LogonType: The type of logon performed.  
ProcessName: The the executable of the process which initiated the logon.  
RestrictedAdminMode: Whether or not the logon was running in restricted admin mode.  
SubjectDomainName: The domain of the account initiating the logon.  
SubjectUserName: The username of the account initiating the logon.   
SubjectUserSid: The security identifer of the account initiating the logon.  
TargetDomainName: The domain the account being logged onto resides in.  
TargetOutboundDomainName: The domain for the user account being used for remote connections, in case of logon type 9.  
TargetOutboundUserName: The username of the account being used for remote connections.  
TargetUserName: The username of the account being logged onto.  
TargetUserSid: The security identier of the account being logged onto.  
WorkstationName: The computer name of the workstation from where the logon originated.  

print(df.columns.tolist())

df.isnull().sum()

#### Finding Columns similar to Lanl dataset  
// To Confirm

Lanl is another AD dataset for benign behavior that we might consider merging with - See 2.0-vb-initial-lanl-benign-data-exploration.ipynb  

'AuthenticationPackage',    
'Destination',  
'DomainName',    
'EventID',    
'FailureReason',  
'LogHost',   
'LogonID',  
'LogonType',  
'LogonTypeDescription',  
'ParentProcessID',   
'ParentProcessName',  
'ProcessID',  
'ProcessName',  
'ServiceName',  
'Source',  
'Status',  
'SubjectDomainName',  
'SubjectLogonID',  
'SubjectUserName',  
'Time',  
'UserName'  

df_lanl = df[['EventID','EventTime','LogonId', 'host', 'Hostname', 'LogonType' ,'SubjectUserName', 'SubjectDomainName',
    'SubjectLogonId','Status', 'SourceName', 'ServiceName',"ServiceSid",'DestinationHostname', 'AuthenticationPackageName',
    'FailureReason','ProcessName', 'ProcessId', 'ParentProcessName','ParentProcessId'
   ]]

df_lanl.isnull().all(0)

df_lanl

for col in list(df_lanl):
    print(col)
    print(df_lanl[col].unique())

df_trimmed = df[['@timestamp',
      'SourceAddress',
      'RemoteMachineID',
      'EventTime',
      'EventID', 
#       'Category',
      'EventType',
      'UserID',
      'AccountName',
#       'Domain', 
#       'RuleName',
#       'Description',
#       'NewProcessName', 
#       'NewProcessId',
      'LmPackageName',
      'Service',
      'LogonProcessName', 
      'WorkstationName',
      'LogonType',
      'EventCountTotal',
      'Workstation',
#       'FailureReason', 
#       'Status',
#       'ServiceName', 
#       'ServiceSid', 
#       'TicketOptions'      
     ]]

for col in list(df_trimmed):
    print(col)
    print(df_trimmed[col].unique())

df['Time'] = pd.to_datetime(df['EventTime']) - pd.DateOffset(years = 50, months = 9 , days=22) 

df.count()

len(df[(df.EventID == 4672)])

path = "../data/raw/purplesharp_ad_playbook_I_2020-10-22042947.json"
df = json.read_json(path, lines=True)

event_list = [1, 3, 8, 10, 11, 12, 4624, 4625, 4648, 4658, 4661, 4663, 4672,4698,4768,5140,5145, 5156, 5158]
df[(df.EventID.isin(event_list))]['EventID'].value_counts()
# df['EventID'].value_counts()

path = "../data/raw/empire_uac_shellapi_fodhelper_2020-09-04032946.json"
df = json.read_json(path, lines=True)

df[(df.EventID.isin(event_list))]['EventID'].value_counts()
#df['EventID'].value_counts()

path = "../data/raw/purplesharp_pe_injection_createremotethread_2020-10-2303121085.json"
df = json.read_json(path, lines=True)

df[(df.EventID.isin(event_list))]['EventID'].value_counts()
#df['EventID'].value_counts()

path = "../data/raw/schtask_create_2020-12-1907003032.json"
df = json.read_json(path, lines=True)

df[(df.EventID.isin(event_list))]['EventID'].value_counts()
#df['EventID'].value_counts()
