==> notebooks/2.0-vb-initial-lanl-benign-data-exploration.ipynb <==
import pyspark
import pandas as pd

from pyspark import SparkContext, SparkConf
from pyspark.sql import SparkSession
spark = SparkSession.builder \
    .master('local[*]') \
    .config("spark.driver.memory", "8g") \
    .appName('vb-app') \
    .getOrCreate()

path = "../data/raw/wls_day-01"
df = spark.read.json(path)

from pyspark.sql.functions import desc
df.groupBy('LogHost').count().orderBy(desc('count')).show()

df.columns

df.select('DomainName').show()

import datetime
import time

file_name = 'lanl.csv' 
df.select('EventID',"ProcessName","Time").toPandas().to_csv(file_name)

# df.groupBy('EventID').count().show()

df.groupBy('UserName').count().show()

# df.groupBy('LogonType').count().show()

# df.groupBy('Source').count().show()

## Convert time into timestamp

# df.groupBy('Time').count().show()

# from pyspark.sql.functions import from_unixtime

# df = df.withColumn('timestamp', from_unixtime((df.Time.cast('bigint')/1000)).cast('timestamp'))


df.take(10)

# df.groupBy('timestamp').count().show()

## Create Window Functions (Fix Usernames)

### Total events per user in an (last) hour window interval  
Ref: https://stackoverflow.com/questions/45806194/pyspark-rolling-average-using-timeseries-data

from pyspark.sql.window import Window
from pyspark.sql import functions as F
new_df = df.select("Time", "UserName", "EventID")
w = (Window.partitionBy('UserName').orderBy(F.col("Time").cast('long')).rangeBetween(-3600, Window.currentRow))
new_df = new_df.withColumn('total_events', F.count("Time").over(w))

new_df.orderBy("Time").take(10)

## Paritition By UserName and Event

w = (Window.partitionBy('UserName', 'EventID').orderBy(F.col("Time").cast('long')).rangeBetween(-3600, Window.currentRow))
new_df = new_df.withColumn('total_per_event', F.count("Time").over(w))

### Total number of Event 4624 per user in an hour window interval  (Successful Logon)


new_df.filter((new_df.UserName == 'Comp502166$') & (new_df.EventID == 4624)).show()

### Total number of Event 4625 per user in an hour window interval  (Failed Logon)

### Total number of Event 4627 per user in an hour window interval  (Group Membership)


### Total number of Event 4648 per user in an hour window interval  (A logon was attempted using explicit credentials)

### Total number of Event 4658 per user in an hour window interval  (handle to an object was closed)

### Total number of Event 4661 per user in an hour window interval  (handle to an object was requsted)

### Total number of Event 4672 per user in an hour window interval  (Special privileges assigned to new logon)

### Total number of Event 4697 per user in an hour window interval  (A service was installed in the system)

### Total number of Event 4698 per user in an hour window interval  (A scheduled task was created)

### Total number of Event 4768 per user in an hour window interval  (Kerberos Authentication)

### Total number of Event 4779 per user in an hour window interval  (A security-enabled local group membership was enumerated)

### Total number of Event 5140 per user in an hour window interval  (A network share object was accessed)

### Total number of Event 5145 per user in an hour window interval  (A network share object was checked to see whether client can be granted desired access)

### Total number of Event 5158 per user in an hour window interval  (The Windows Filtering Platform has permitted a bind to a local port)

## Concatenate datasets
