==> notebooks/2.1-vb-test-rolling-window.ipynb <==
import findspark
findspark.init()
import pyspark

# from pyspark import SparkContext, SparkConf
# from pyspark.sql import Window
# from pyspark.sql import functions as func
# from pyspark.sql import SQLContext
# sc = SparkContext.getOrCreate()
# # sqlContext = SQLContext(sc)
# # tup = [(26, "a"), (29, "a"), (374, "a"), (6050, "b"), (6989, "b"), (8888, "b")]
# # df = sqlContext.createDataFrame(tup, ["id", "category"])
# # df.show()

# window = Window.partitionBy("category").orderBy("id").rangeBetween(-3600, 0)
# df.withColumn("sum", func.count("id").over(window)).sort("category").show()

from pyspark.sql import SparkSession
spark = SparkSession.builder \
    .master('local[*]') \
    .config("spark.driver.memory", "8g") \
    .appName('vb-app') \
    .getOrCreate()



from pyspark.sql.functions import udf
from pyspark.sql.types import IntegerType

path = "../data/raw/purplesharp_ad_playbook_I_2020-10-22042947.json"
df = spark.read.json(path)
def convert_to_unix(event_time):
    print(event_time)
    time = event_time[10:]
    print(time)
    time_list = time.split(":")
    return int(time_list[0])*3600+int(time_list[1])*60+int(time_list[2])



convert_to_unix_udf = udf(convert_to_unix, IntegerType())

df = df.withColumn("Time", convert_to_unix_udf(df.EventTime))
df.select("Time").show()

df.select('Time').distinct().collect()
