==> notebooks/3.0-ss-initial-atp-simulator-cobalt-strike-data-exploration.ipynb <==
**Import modules**

import pandas as pd
import datetime
import time
from pandas.io import json
import pyspark
from pyspark import SparkContext, SparkConf, SQLContext
from pyspark.sql import SparkSession
from pyspark.sql.functions import unix_timestamp
from pyspark.sql.types import *
conf = pyspark.SparkConf().setAppName('appName').setMaster('local')
sc = SparkContext(conf=conf)
spark = SparkSession(sc)
spark.conf.set("spark.sql.caseSensitive", "true")
sqlContext = SQLContext(SparkContext.getOrCreate())
spark.conf.set("spark.sql.execution.arrow.enabled", "true")

**Import dataset**

path = "../data/raw/aptsimulator_cobaltstrike_2021-06-11T21081492.json"
df = pd.read_json(path,lines=True)
df2 = spark.read.json(path)

df.isnull().sum()

**Initial view of the the dataset**

df.head()

df['TimeCreated'] = pd.to_datetime(df['TimeCreated']) - pd.DateOffset(years = 51, months = 5, days=10) 
df.head()

def timetounix(t):
    return (time.mktime(datetime.datetime.strptime(str(t), "%Y-%m-%d %H:%M:%S.%f000+00:00").timetuple()))

df['Time'] = df['TimeCreated'].apply(timetounix)

df['Time']

df = df[['EventID',"ProcessName","Time"]]

df.head()
file_name = 'apt.csv'
df.to_csv(file_name)

#### Notes:

- Work on discretizing, normalizing and think of how to create features from the columns and have questions about the data

#### Findings:

- Trying to find columns similar to LANL Denign AD Logs & PurpleSharp AD Log Playbook for Lateral Movement

'SourceName', <br>
'Hostname', <br>
'TimeCreated', <-- 'EventTime' (LANL) <br>
'DestinationIsIpv6', 'DestinationIp', 'DestinationHostname', 'DestinationPort', 'DestinationPortName'; <-- (Look into these) <br>
'SubjectDomainName', <br>
'EventID', <br>
'EventType', <br>
'LogonId', <br>
'ParentProcessId', <br>
'ParentProcessName', <br>
'ProcessID', <br>
'ProcessName', <br>
'NewProcessId', <br>
'NewProcessName', <br>
'ServiceName', <br>
'SubjectUserSid', <br>
'SubjectUserName', <br>
'SubjectDomainName', <br>
'SubjectLogonId', <br>
'Status' <br>

'AccessReason', 'Task', 'Message', 'Description', 'Details' <-- (Look into these)

df_cobalt = df[['EventID', 'EventType', 'LogonId', 'Hostname', 'SubjectUserName', 'SubjectDomainName', 'SubjectLogonId', 
                'Status', 'SourceName', 'ServiceName', 'DestinationHostname', 'ProcessName', 'ProcessID', 'ParentProcessName',
                'ParentProcessId']]

df_cobalt.head()

df_cobalt.isnull().all(0)

for col in list(df_cobalt):
    print(col)
    print(df_cobalt[col].unique())

df_cobalt.groupby(by=['ServiceName']).get_group('tbbd05')

df2.groupby('Status').count().show()

df2.groupby('Task').count().show()

df2.groupby('SourceName').count().show()

df2.groupby('Hostname').count().show()

df2.groupby('EventID').count().show()

df2.groupby('LogonId').count().show()                

df2.groupby('Channel').count().show()

df2.groupby('Keywords').count().show()

df2.groupby('Message').count().show()

df2.groupby('SourceName','EventID').count().show()

df2.groupby('SourceName','LogonId').count().show()
