==> notebooks/3.1-ss-test-pivot-standardize.ipynb <==
import findspark
findspark.init()
import pyspark
from pyspark import SparkContext, SparkConf, SQLContext
from pyspark.sql import SparkSession
from pyspark.sql.functions import unix_timestamp
from pyspark.sql.types import *
conf = pyspark.SparkConf().setAppName('appName').setMaster('local')
sc = SparkContext(conf=conf)
spark = SparkSession(sc)

path = "../data/processed/benign_with_rolling_window.csv"
df = spark.read.csv(path).toDF(
  "time", "username", "event_id", "total_events", "total_per_event"
)
df.show()

keep_event_ids = [4624, 4625, 4627, 4648, 4658, 4661, 4672, 4697, 4698, 4768, 4779, 5140, 5145, 5158]

df.printSchema()

df = df.withColumn("event_id", df["event_id"].cast(IntegerType()))
df = df.withColumn("total_events", df["total_events"].cast(IntegerType()))
df = df.withColumn("total_per_event", df["total_per_event"].cast(IntegerType()))

df.printSchema()

df = df.filter(df.event_id.isin(keep_event_ids))
df.show()

# Pivot:
df = df.groupBy("time", "username", "total_events").pivot("event_id").sum("total_per_event")
df.show()

df.printSchema()

df = df.na.fill(value=0, subset=["4624", "4625", "4648", "4672", "4768"])
df.show()

split_df = df.filter(df.username.isin("User793535", "User031784", "User318330", "User974844"))
split_df.show()

# split_df.toPandas().to_csv('mycsv.csv')
# split_df.write.option("header",True).csv("Desktop/Team Elysium/anomaly_detection_active_directory_logs/data/processed/pivotdata.csv")

# Standardize:
from pyspark.ml import Pipeline
from pyspark.ml.feature import MinMaxScaler, VectorAssembler
from pyspark.sql.functions import udf
from pyspark.sql.types import DoubleType

unlist = udf(lambda x: round(float(list(x)[0]),3), DoubleType())

# columns_to_scale = ["total_events", "4624", "4625", "4648", "4672", "4768"]

# assemblers = [VectorAssembler(inputCols=[col], outputCol=col + "_vec") for col in columns_to_scale]

# scalers = [MinMaxScaler(inputCol=col + "_vec", outputCol=col + "_scaled") for col in columns_to_scale]
# pipeline = Pipeline(stages=assemblers + scalers)
# scalerModel = pipeline.fit(df)
# scaledData = scalerModel.transform(df)

for i in ["total_events", "4624", "4625", "4648", "4672", "4768"]:
    assembler = VectorAssembler(inputCols=[i], outputCol=i+"_Vect")
    scaler = MinMaxScaler(inputCol=i+"_Vect", outputCol=i+"_Scaled")
    pipeline = Pipeline(stages=[assembler, scaler])
    split_df = pipeline.fit(split_df).transform(split_df).withColumn(i+"_Scaled", unlist(i+"_Scaled")).drop(i+"_Vect")

print("After Scaling:")
split_df.show()
