==> notebooks/4.0-vr-test-concatenation.ipynb <==
**Import modules**

import pandas as pd
from pandas import DataFrame

apt = pd.read_csv('../data/interim/apt.csv')
lanl = pd.read_csv('../data/interim/lanl.csv')
purple = pd.read_csv('../data/interim/purple.csv')

apt = apt.drop(apt.columns[0],axis=1)
purple = purple.drop(purple.columns[0],axis=1)
lanl = lanl.drop(lanl.columns[0],axis=1)

apt['Anomalous'] = 1
purple['Anomalous'] = 1
lanl['Anomalous'] = 0

apt.head()

purple.head()

lanl.head()

datasets = [apt, purple, lanl]

finaldf = pd.concat(datasets)

finaldf.head()

finaldf = finaldf.sort_values(by=['Time'])

finaldf.isnull().sum()

filename = 'concatenated.csv'

finaldf.to_csv(filename)

import pyspark
from pyspark import SparkContext, SparkConf, SQLContext
from pyspark.sql import SparkSession
from pyspark.sql.functions import unix_timestamp
from pyspark.sql.types import *
conf = pyspark.SparkConf().setAppName('appName').setMaster('local')
sc = SparkContext(conf=conf)
spark = SparkSession(sc)


df = spark.read.csv("concatenated.csv")
df.printSchema()

df.show()
