==> notebooks/7.0-vb-test-production-model.ipynb <==
from pickle import dump, load
import pandas as pd
from sklearn.ensemble import RandomForestClassifier
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split
from sklearn.pipeline import Pipeline
from imblearn.over_sampling import SMOTE
from sklearn.metrics import  classification_report

df = pd.read_csv("../data/processed/final_df.csv")
columns = list(df.columns)
features = columns[4:]
df[features] = df[features].div(df.total_events, axis = 0)
# Separating out the features
X = df.loc[:, features].values
y = df.loc[:, ['is_malicious']].values
print(X)
print(y)

smote = SMOTE(random_state=0, sampling_strategy="minority")
X_os, y_os = smote.fit_resample(X, y)
X_train_os, X_test_os, y_train_os, y_test_os = train_test_split(X_os, y_os, test_size = 0.2, random_state=2)

model = Pipeline([('scaler', StandardScaler()), ('forest', RandomForestClassifier())])
model.fit(X_train_os, y_train_os)

y_pred_os = model.predict(X_test_os)
print(classification_report(y_test_os, y_pred_os))

with open('../src/models/final_model.sav', 'wb') as f:
    dump(model, f)

with open('../src/models/final_model.sav', 'rb') as f:
    loaded_model = load(f)

y_pred_os = loaded_model.predict(X_test_os)
print(classification_report(y_test_os, y_pred_os))

df[df.is_malicious == 1]

df = pd.read_csv("../data/processed/final_df.csv")
columns = list(df.columns)
features = columns[4:]
df[features] = df[features].div(df.total_events, axis = 0)
data = df.iloc[104,4:].values
data

loaded_model.predict(data.reshape(1,-1))[0]
